Remove unused parameters (#1685)
* Remove unused parameters

* Remove unnecessary stded25519.PrivateKey conversion

* Use defer to cleanup workerpool tasks

* Fix missing or wrong error checks

* Remove unnecessary type cast

* Add context to snapshot read functions

* Add verification of ms payload in verifyDatabase

* Fix linter warnings in "random" util package

* Fix json tags in toolset outputs

* Fix json tags in coordinator state file

* Fix revive linter warnings

* Add nolint directives to test files

* Add whitespace to comments

* Fix review comments

+++ pkg/toolset/snap_gen.go
@@ -296,7 +296,7 @@ func snapshotGen(args []string) error {
 			genesisBalancesIndex++
 
 			return utxo.CreateOutput(
-				iotago.OutputIDFromTransactionIDAndIndex(TransactionIDFromIndex(int64(genesisBalancesIndex)), 0),
+				iotago.OutputIDFromTransactionIDAndIndex(TransactionIDFromIndex(genesisBalancesIndex), 0),
 				iotago.EmptyBlockID(),
 				0,
 				0,
